code changes for share leads

+++ src/app/contacts/add-contacts/add-contacts.component.ts
@@ -806,7 +806,7 @@ export class AddContactsComponent implements OnInit, OnDestroy {
     oneAtTimeSaveAfterGotPermition() {
 
         if (this.assignLeads) {
-            this.contactListObject.contactType="ASSIGNED_LEADS_LIST";
+            this.contactListObject.contactType="CONTACT";
             this.contactListObject.socialNetwork="MANUAL";
             this.userUserListWrapper.users = this.newUsers;
             this.saveAssignedLeadsList();
@@ -980,7 +980,7 @@ export class AddContactsComponent implements OnInit, OnDestroy {
 
     saveClipBoardContactsAfterGotPermition() {
         if (this.assignLeads) {
-            this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+            this.contactListObject.contactType = "CONTACT";
             this.contactListObject.socialNetwork = "MANUAL";
             this.setLegalBasisOptions(this.clipboardUsers);
             this.userUserListWrapper.users = this.clipboardUsers;
@@ -1111,7 +1111,7 @@ export class AddContactsComponent implements OnInit, OnDestroy {
 
     saveCsvContactsWithPermission() {
            if (this.assignLeads) {
-            this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+            this.contactListObject.contactType = "CONTACT";
             this.contactListObject.socialNetwork = "MANUAL";
             this.setLegalBasisOptions(this.contacts);
             this.userUserListWrapper.users = this.contacts;
@@ -1632,7 +1632,7 @@ export class AddContactsComponent implements OnInit, OnDestroy {
             this.contactListObject = new ContactList;
             this.contactListObject.name = this.model.contactListName;
             this.contactListObject.isPartnerUserList = this.isPartner;
-            this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+            this.contactListObject.contactType = "CONTACT";
             this.contactListObject.socialNetwork = this.socialContact.socialNetwork;
             this.contactListObject.publicList = true;
             this.contactListObject.synchronisedList = true;
@@ -1710,7 +1710,7 @@ export class AddContactsComponent implements OnInit, OnDestroy {
                this.contactListObject = new ContactList;
                this.contactListObject.name = this.model.contactListName;
                this.contactListObject.isPartnerUserList = this.isPartner;
-               this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+               this.contactListObject.contactType = "CONTACT";
                this.contactListObject.socialNetwork = "MANUAL";
                this.contactListObject.publicList = true;
                this.setLegalBasisOptions(this.allselectedUsers);
@@ -2069,7 +2069,7 @@ export class AddContactsComponent implements OnInit, OnDestroy {
         	      this.contactListObject = new ContactList;
                   this.contactListObject.name = this.model.contactListName;
                   this.contactListObject.isPartnerUserList = this.isPartner;
-                  this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+                  //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
                   this.contactListObject.synchronisedList = true;
                   this.contactListObject.socialNetwork = this.socialContact.socialNetwork;
                   this.contactListObject.publicList = true;
@@ -2150,7 +2150,7 @@ export class AddContactsComponent implements OnInit, OnDestroy {
             this.contactListObject = new ContactList;
             this.contactListObject.name = this.model.contactListName;
             this.contactListObject.isPartnerUserList = this.isPartner;
-            this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+            //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
             this.contactListObject.socialNetwork = "MANUAL";
             this.contactListObject.publicList = true;
             this.setLegalBasisOptions(this.allselectedUsers);
@@ -2546,7 +2546,7 @@ salesForceVanityAuthentication() {
         	     this.contactListObject = new ContactList;
                  this.contactListObject.name = this.model.contactListName;
                  this.contactListObject.isPartnerUserList = this.isPartner;
-                 this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+                 //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
                  this.contactListObject.socialNetwork = "MANUAL";
                  this.contactListObject.publicList = true;
                  this.setLegalBasisOptions(this.allselectedUsers);
@@ -2624,7 +2624,7 @@ salesForceVanityAuthentication() {
             this.contactListObject = new ContactList;
             this.contactListObject.name = this.model.contactListName;
             this.contactListObject.isPartnerUserList = this.isPartner;
-            this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+            //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
             this.contactListObject.synchronisedList = true;
             this.contactListObject.socialNetwork = this.socialContact.socialNetwork;
             this.contactListObject.publicList = true;
@@ -2998,7 +2998,7 @@ salesForceVanityAuthentication() {
 				}
 				else if (e.data == 'isZohoAuth') {
 					localStorage.setItem('isZohoAuth', 'yes');
-				}else if(e.data !=''){
+				}else if(e.data !=null && e.data.includes("You have already configured")){
 					localStorage.setItem('validationMessage', e.data);
 				}
 			}, false);
@@ -3291,7 +3291,7 @@ vanityCheckingMarketoContactsAuthentication(){
                this.contactListObject = new ContactList;
                this.contactListObject.name = this.model.contactListName;
                this.contactListObject.isPartnerUserList = this.isPartner;
-               this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+               //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
                this.contactListObject.synchronisedList = true;
                this.contactListObject.socialNetwork = this.socialContact.socialNetwork;
                this.contactListObject.publicList = true;
@@ -3362,7 +3362,7 @@ vanityCheckingMarketoContactsAuthentication(){
         	     this.contactListObject = new ContactList;
                  this.contactListObject.name = this.model.contactListName;
                  this.contactListObject.isPartnerUserList = this.isPartner;
-                 this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+                 //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
                  this.contactListObject.socialNetwork = "MANUAL";
                  this.contactListObject.publicList = true;
                  this.setLegalBasisOptions(this.allselectedUsers);
@@ -3888,7 +3888,7 @@ vanityCheckingMarketoContactsAuthentication(){
                this.contactListObject = new ContactList;
                this.contactListObject.name = this.model.contactListName;
                this.contactListObject.isPartnerUserList = this.isPartner;
-               this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+               //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
                this.contactListObject.synchronisedList = true;
                this.contactListObject.socialNetwork = this.socialContact.socialNetwork;
                this.contactListObject.publicList = true;
@@ -3935,7 +3935,7 @@ vanityCheckingMarketoContactsAuthentication(){
             this.contactListObject = new ContactList;
             this.contactListObject.name = this.model.contactListName;
             this.contactListObject.isPartnerUserList = this.isPartner;
-            this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
+            //this.contactListObject.contactType = "ASSIGNED_LEADS_LIST";
             this.contactListObject.socialNetwork = "MANUAL";
             this.contactListObject.publicList = true;
             this.setLegalBasisOptions(this.allselectedUsers);
